replace old raster pipe
