+++ gfw_pixetl/pipes/pipe_factory.py
@@ -1,18 +1,14 @@
 from typing import List, Optional
 
 from gfw_pixetl.layers import RasterSrcLayer, VectorSrcLayer, Layer
-from gfw_pixetl.pipes import Pipe, VectorPipe, RasterPipe, CalcRasterPipe
+from gfw_pixetl.pipes import Pipe, VectorPipe, RasterPipe
 
 
-def pipe_factory(
-    layer: Layer, subset: Optional[List[str]] = None, divisor: int = 2
-) -> Pipe:
+def pipe_factory(layer: Layer, subset: Optional[List[str]] = None) -> Pipe:
     if isinstance(layer, VectorSrcLayer):
-        pipe: Pipe = VectorPipe(layer, subset, divisor)
-    elif isinstance(layer, RasterSrcLayer) and layer.calc is not None:
-        pipe = CalcRasterPipe(layer, subset, divisor)
+        pipe: Pipe = VectorPipe(layer, subset)
     elif isinstance(layer, RasterSrcLayer):
-        pipe = RasterPipe(layer, subset, divisor)
+        pipe = RasterPipe(layer, subset)
     else:
         raise ValueError("Unknown layer type")
