+++ tests/test_raster_pipe.py
@@ -31,54 +31,44 @@ PIPE = RasterPipe(LAYER, SUBSET)
 def test_create_tiles_subset():
     with mock.patch.object(RasterSrcTile, "src_tile_intersects", return_value=True):
         with mock.patch.object(RasterSrcTile, "dst_exists", return_value=False):
-            with mock.patch.object(RasterSrcTile, "transform", return_value=None):
-                with mock.patch.object(
-                    RasterSrcTile, "local_src_is_empty", return_value=False
-                ):
+            with mock.patch.object(RasterSrcTile, "transform", return_value=True):
+                # with mock.patch.object(
+                #         RasterSrcTile, "local_src_is_empty", return_value=False
+                # ):
+                with mock.patch.object(RasterSrcTile, "upload", return_value=None):
                     with mock.patch.object(
-                        RasterSrcTile, "compress", return_value=None
+                        RasterSrcTile, "rm_local_src", return_value=None
                     ):
                         with mock.patch.object(
-                            RasterSrcTile, "upload", return_value=None
+                            RasterPipe, "upload_vrt", return_value=None
                         ):
                             with mock.patch.object(
-                                RasterSrcTile, "rm_local_src", return_value=None
+                                RasterPipe, "upload_extent", return_value=None
                             ):
-                                with mock.patch.object(
-                                    RasterPipe, "upload_vrt", return_value=None
-                                ):
-                                    with mock.patch.object(
-                                        RasterPipe, "upload_extent", return_value=None
-                                    ):
-                                        result = PIPE.create_tiles()
-                                        assert len(result) == 3
+                                result = PIPE.create_tiles()
+                                assert len(result) == 3
 
 
 def test_create_tiles_all():
     pipe = RasterPipe(LAYER)
     with mock.patch.object(RasterSrcTile, "src_tile_intersects", return_value=True):
         with mock.patch.object(RasterSrcTile, "dst_exists", return_value=False):
-            with mock.patch.object(RasterSrcTile, "transform", return_value=None):
-                with mock.patch.object(
-                    RasterSrcTile, "local_src_is_empty", return_value=False
-                ):
+            with mock.patch.object(RasterSrcTile, "transform", return_value=True):
+                # with mock.patch.object(
+                #         RasterSrcTile, "local_src_is_empty", return_value=False
+                # ):
+                with mock.patch.object(RasterSrcTile, "upload", return_value=None):
                     with mock.patch.object(
-                        RasterSrcTile, "compress", return_value=None
+                        RasterSrcTile, "rm_local_src", return_value=None
                     ):
                         with mock.patch.object(
-                            RasterSrcTile, "upload", return_value=None
+                            RasterPipe, "upload_vrt", return_value=None
                         ):
                             with mock.patch.object(
-                                RasterSrcTile, "rm_local_src", return_value=None
+                                RasterPipe, "upload_extent", return_value=None
                             ):
-                                with mock.patch.object(
-                                    RasterPipe, "upload_vrt", return_value=None
-                                ):
-                                    with mock.patch.object(
-                                        RasterPipe, "upload_extent", return_value=None
-                                    ):
-                                        result = pipe.create_tiles()
-                                        assert len(result) == 648
+                                result = pipe.create_tiles()
+                                assert len(result) == 648
 
 
 def test_filter_src_tiles():
@@ -102,7 +92,7 @@ def test_filter_src_tiles():
 
 
 def test_transform():
-    with mock.patch.object(RasterSrcTile, "transform", return_value=None):
+    with mock.patch.object(RasterSrcTile, "transform", return_value=True):
         tiles = _get_subset_tiles()
         tiles = PIPE.transform(tiles)
         i = 0
@@ -112,15 +102,7 @@ def test_transform():
         assert i == 4
 
 
-def test_compress():
-    with mock.patch.object(RasterSrcTile, "compress", return_value=None):
-        tiles = _get_subset_tiles()
-        tiles = PIPE.compress(tiles)
-        i = 0
-        for tile in tiles:
-            i += 1
-            assert isinstance(tile, RasterSrcTile)
-        assert i == 4
+# ./s
 
 
 def _get_subset_tiles() -> Set[RasterSrcTile]:
